+++ cli/package.json
@@ -9,6 +9,7 @@
 	"dependencies": {
 		"@plexusjs/core": "^1.7.4",
 		"chalk": "^5.0.1",
+		"glob": "^10.2.6",
 		"yargs": "^17.4.1"
 	},
 	"scripts": {
